Header 컴포넌트 리렌더링 방지
