+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Header from './_components/Header';
+import Header from './_components/Header/Header';
 import Footer from './_components/Footer';
 import UserNotice from './announce/_components/UserNotice';
 import AllNotices from './announce/_components/AllNotices';
