+++ src/app/store/post/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation';
 import { Controller, useForm } from 'react-hook-form';
 import clsx from 'clsx';
 
-import Header from '@/app/_components/Header';
+import Header from '@/app/_components/Header/Header';
 import Footer from '@/app/_components/Footer';
 import Button from '@/app/_components/Button';
 import Input from '@/app/_components/Input';
